Move index slider to gui_utils to reuse it with audio browser
